+++ src/main/scala/elaio/neuralnet/training/WeightInitializer.scala
@@ -7,7 +7,7 @@ object WeightInitializer {
   // He-style init, inverted because collectInConnections averages rather than sums:
   // Var(w) = N / secondMomentFactor, so weights grow with fan-in instead of shrinking.
   // Runs after the graph is built, walking back from the outputs once fan-in is final.
-  def initialize(outputNodes: Array[Neuron]): Int = {
+  def initialize(outputNodes: Array[Neuron]): Double = {
     val random = new scala.util.Random
     var connectionsInitialized = 0
